finish project

+++ bitboard.go
@@ -12,36 +12,32 @@ func (b bitBoard) count() int {
 	return bits.OnesCount64(uint64(b))
 }
 
-func (b *bitBoard) set(pos uint) {
-	*b |= bitBoard(uint64(1) << pos)
+func (b *bitBoard) set(pos int) {
+	*b |= bitBoard(uint64(1) << uint(pos))
 }
 
-func (b bitBoard) test(pos uint) bool {
-	return (b & bitBoard(uint64(1)<<pos)) != 0
+func (b bitBoard) test(pos int) bool {
+	return (b & bitBoard(uint64(1)<<uint(pos))) != 0
 }
 
-func (b *bitBoard) clr(pos uint) {
-	*b &= bitBoard(^(uint64(1) << pos))
+func (b *bitBoard) clr(pos int) {
+	*b &= bitBoard(^(uint64(1) << uint(pos)))
 }
 
 func (b *bitBoard) firstOne() int {
 	bit := bits.TrailingZeros64(uint64(*b))
-
 	if bit == 64 {
 		return 64
 	}
-
 	*b = (*b >> uint(bit+1)) << uint(bit+1)
 	return bit
 }
 
-// TODO: lastOne() here
 func (b *bitBoard) lastOne() int {
 	bit := bits.LeadingZeros64(uint64(*b))
 	if bit == 64 {
 		return 64
 	}
-
 	*b = (*b << uint(bit+1)) >> uint(bit+1)
 	return 63 - bit
 }
@@ -88,3 +84,4 @@ func (b bitBoard) Stringln() string {
 	s = strings.Replace(s, "0", "0 ", -1)
 	return s
 }
+
